server/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
pypdf
python-docx
chromadb
requests
google-generativeai
PyMuPDF
easyocr
Pillow